Move feature detection functions to new module
